+++ pac4j-core/src/main/java/org/pac4j/core/http/ajax/AjaxRequestResolver.java
@@ -1,6 +1,7 @@
 package org.pac4j.core.http.ajax;
 
 import org.pac4j.core.context.WebContext;
+import org.pac4j.core.redirect.RedirectAction;
 
 /**
  * Compute if a HTTP request is an AJAX one.
@@ -17,4 +18,13 @@ public interface AjaxRequestResolver {
      * @return whether it is an AJAX request
      */
     boolean isAjax(WebContext context);
+
+    /**
+     * Build an AJAX reponse.
+     *
+     * @param url the redirection URL if it was not an AJAX request
+     * @param context the web context
+     * @return the AJAX response
+     */
+    RedirectAction buildAjaxResponse(String url, WebContext context);
 }